common shape: handling dash patterns < 1
For the dash pattern values (both, length and gap) < 1 the program crashed.
curLen - variable type changed from int to float.

+++ src/lib/sw_engine/tvgSwCommon.h
@@ -165,7 +165,7 @@ struct SwStroke
 struct SwDashStroke
 {
     SwOutline* outline;
-    int32_t curLen;
+    float curLen;
     int32_t curIdx;
     Point ptStart;
     Point ptCur;
